requirements.txt
absl-py==0.15.0
apache-beam==2.38.0
astunparse==1.6.3
avro-python3==1.10.2
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.3
cloudpickle==2.0.0
colorama==0.4.4
contextlib2==21.6.0
crcmod==1.7
cycler==0.11.0
Cython==0.29.28
dill==0.3.1.1
dm-tree==0.1.7
docopt==0.6.2
fastavro==1.4.11
Flask==2.1.2
flatbuffers==1.12
fonttools==4.33.3
gast==0.4.0
gin-config==0.5.0
google-api-core==2.7.3
google-api-python-client==2.47.0
google-auth==2.6.6
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.56.0
grpcio==1.34.1
h5py==3.1.0
hdfs==2.7.0
httplib2==0.19.1
idna==3.3
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.1.0
kaggle==1.5.12
keras==2.8.0
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
kiwisolver==1.4.2
libclang==14.0.1
lvis==0.5.3
lxml==4.8.0
Markdown==3.3.7
MarkupSafe==2.1.1
matplotlib==3.5.2
numpy==1.22.3
oauth2client==4.1.3
oauthlib==3.2.0
object-detection @ file:///D:/startPython/SrartFalsk/Neuai/TensorFlow/models/research
opencv-python==4.5.5.64
opencv-python-headless==4.5.5.64
opt-einsum==3.3.0
orjson==3.6.8
packaging==21.3
pandas==1.4.2
Pillow==9.1.0
portalocker==2.4.0
promise==2.3
proto-plus==1.20.3
protobuf==3.20.1
psutil==5.9.0
py-cpuinfo==8.0.0
pyarrow==6.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools @ git+https://github.com/philferriere/cocoapi.git@2929bd2ef6b451054755dfd7ceb09278f935f7ad#subdirectory=PythonAPI
pydot==1.4.2
pymongo==3.12.3
pyparsing==2.4.7
python-dateutil==2.8.2
python-slugify==6.1.2
pytz==2022.1
pywin32==304
PyYAML==5.4.1
regex==2022.4.24
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
sacrebleu==2.0.0
scikit-learn==1.0.2
scipy==1.8.0
sentencepiece==0.1.96
seqeval==1.2.2
six==1.15.0
tabulate==0.8.9
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.8.0
tensorflow-addons==0.16.1
tensorflow-datasets==4.5.2
tensorflow-estimator==2.5.0
tensorflow-hub==0.12.0
tensorflow-io==0.25.0
tensorflow-io-gcs-filesystem==0.25.0
tensorflow-metadata==1.8.0
tensorflow-model-optimization==0.7.2
tensorflow-text==2.8.2
termcolor==1.1.0
text-unidecode==1.3
tf-estimator-nightly==2.8.0.dev2021122109
tf-models-official==2.8.0
tf-slim==1.1.0
threadpoolctl==3.1.0
tqdm==4.64.0
typeguard==2.13.3
typing-extensions==3.7.4.3
uritemplate==4.1.1
urllib3==1.26.9
Werkzeug==2.1.2
wincertstore==0.2
wrapt==1.12.1
zipp==3.8.0
